Revert "Merge branch 'develop' into y24-432-centralize-vue-component-initialization"
This reverts commit ed4e91f712a011ec010153d41f60982de68db238, reversing
changes made to 94661961ed07670bb59528c3152c788754b0e177.
