+++ app/frontend/stylesheets/limber/screen.scss
@@ -269,12 +269,6 @@ nav.robots-list {
   @extend .mb-2;
 }
 
-.choose-workflow {
-  .create-submission-button {
-    @extend .btn-success;
-  }
-}
-
 .other-actions {
   .work-completion-button,
   .create-plate-button,
@@ -285,7 +279,8 @@ nav.robots-list {
 .suggested-actions {
   .work-completion-button,
   .create-plate-button,
-  .create-tube-button {
+  .create-tube-button,
+  .create-submission-button {
     @extend .btn-success;
   }
 }
