+++ app/models/presenters/submission_plate_presenter.rb
@@ -17,12 +17,8 @@ module Presenters
   class SubmissionPlatePresenter < PlatePresenter
     include Presenters::Statemachine::Submission
     include Presenters::Statemachine::DoesNotAllowLibraryPassing
-    include Presenters::SubmissionBehaviour
     include Presenters::StateChangeless
 
-    self.allow_well_failure_in_states = []
-
-    # Stock style class causes well state to inherit from plate state.
     self.style_class = 'stock'
 
     self.summary_items = {
@@ -33,5 +29,35 @@ module Presenters
       'Input plate barcode' => :input_barcode,
       'Created on' => :created_on
     }
+
+    self.allow_well_failure_in_states = []
+
+    def each_submission_option
+      purpose_config.submission_options.each do |button_text, options|
+        submission_options = options.to_hash
+        submission_options[:asset_groups] = asset_groups
+        submission_options[:labware_barcode] = labware.labware_barcode.human
+        yield button_text, SequencescapeSubmission.new(submission_options)
+      end
+    end
+
+    def pending_submissions?
+      submissions.any? { |submission| submission.building_in_progress?(ready_buffer: 20.seconds) }
+    end
+
+    def submissions
+      labware.direct_submissions
+    end
+
+    private
+
+    def asset_groups
+      @asset_groups ||=
+        labware
+          .wells
+          .compact_blank
+          .group_by(&:order_group)
+          .map { |_, wells| { assets: wells.map(&:uuid), autodetect_studies: true, autodetect_projects: true } }
+    end
   end
 end
