+++ app/views/plates/_choose_workflow.html.erb
@@ -1,9 +1,10 @@
 <%= vite_javascript_tag 'entrypoints/pages/choose_workflow.js', defer: true %>
-<%= card title:'Choose workflow', id: 'choose_workflow_card', css_class: 'choose-workflow logged_in_only' do %>
+<%= card title:'Choose workflow', id: 'choose_workflow_card', css_class: 'suggested-actions logged_in_only' do %>
   <p class="lead">
-    A new workflow can be created for this plate. Please choose a workflow to
-    continue. This will build the necessary submissions in Sequencescape,
-    allowing you to proceed.
+    This plate can be processed via multiple workflows and there are currently
+    no active workflows associated with the plate. Please select a workflow.
+    This will build the necessary submissions in Sequencescape, and will allow
+    you to proceed.
   </p>
 
   <div id="submission_forms">
