+++ app/views/plates/_pending_submissions.html.erb
@@ -1,9 +1,8 @@
-<%# presenter.submissions returns a JsonApiClient::Query::Builder at this point #%>
-<% if presenter.submissions.present? and presenter.submissions.size > 0 %>
-  <%= card title:'Submissions', css_class: 'submissions', without_block: true do %>
+<% if presenter.submissions.present? %>
+  <%= card title:'Submissions', css_class: 'suggested-actions', without_block: true do %>
     <% if presenter.pending_submissions? %>
       <div class="card-body">
-        <p class="lead">There are currently submissions being built, the page will refresh in
+        <p class="lead">There are currently submissions being built, This page will refresh in
         <span id="refresh_time" data-reload-time="5">5</span> seconds.</p>
       </div>
     <% end %>
