--- app/views/plates/sidebars/_container.html.erb
@@ -1,4 +1,0 @@
-<%= sidebar do %>
-  <%= render 'logged_out_warning' %>
-  <%= render "plates/sidebars/#{@presenter.sidebar_partial}", presenter: @presenter %>
-<% end %>
