+++ app/views/plates/sidebars/_default.html.erb
@@ -1,17 +1,33 @@
-<%= render 'plate_printing' %>
-<% presenter.tubes_and_sources do %>
-  <%= render 'tube_printing', presenter: presenter %>
-<% end %>
-
-<%= render 'labware/qc_data_upload' %>
-
-<%= card title:'Suggested Actions', css_class: 'suggested-actions' do %>
-  <% presenter.control_suggested_library_passing do %>
-    <%= render 'work_completion_form', presenter: presenter %>
+<%= sidebar do %>
+  <%= render 'logged_out_warning' %>
+  <%= render 'plate_printing' %>
+  <% presenter.tubes_and_sources do %>
+    <%= render 'tube_printing', presenter: presenter %>
   <% end %>
 
-  <% presenter.control_additional_creation do -%>
-    <%= render presenter.suggested_purposes, parent: presenter.labware %>
+  <%= render 'labware/qc_data_upload' %>
+
+  <%= card title:'Suggested Actions', css_class: 'suggested-actions' do %>
+    <% presenter.control_suggested_library_passing do %>
+      <%= render 'work_completion_form', presenter: presenter %>
+    <% end %>
+
+    <% presenter.control_additional_creation do -%>
+      <%= render presenter.suggested_purposes, parent: presenter.labware %>
+    <% end %>
+
+    <% presenter.csv_file_links.each do |link_name, link_target| %>
+      <%= link_to(link_name, link_target, rel: "external",  class: 'csv-link') %>
+    <% end %>
+
+    <% if presenter.robot? %>
+      <h4 class="card-title mt-3">Robots</h4>
+      <% presenter.each_robot do |robot_id, robot_name| %>
+        <%= link_to robot_name, robot_path(robot_id), class: 'btn btn-primary btn-lg btn-block logged_in_only' %>
+      <% end %>
+    <% else %>
+      <%= simple_state_change_form(presenter) %>
+    <% end %>
   <% end %>
 
   <%= card title: 'Adding a Comment', css_class: 'logged_in_only' do %>
@@ -22,58 +38,43 @@
       data-sequencescape-url="<%= Limber::Application.config.sequencescape_url %>"
       data-asset-id="<%= presenter.id %>">
     </div>
-  <% presenter.csv_file_links.each do |link_name, link_target| %>
-    <%= link_to(link_name, link_target, rel: "external",  class: 'csv-link') %>
   <% end %>
 
-  <% if presenter.robot? %>
-    <h4 class="card-title mt-3">Robots</h4>
-    <% presenter.each_robot do |robot_id, robot_name| %>
-      <%= link_to robot_name, robot_path(robot_id), class: 'btn btn-primary btn-lg btn-block logged_in_only' %>
+  <%= card title: 'Adding Custom Metadata', css_class: 'logged_in_only' do %>
+    <div id="labware-custom-metadata-add-form"
+      data-custom-metadata-fields="<%= presenter.custom_metadata_fields %>"
+      data-sequencescape-api="<%= Limber::Application.config.api.v2.connection_options.js_url %>"
+      data-sequencescape-api-key="<%= Limber::Application.config.api.v2.connection_options.authorisation %>"
+      data-sequencescape-url="<%= Limber::Application.config.sequencescape_url %>"
+      data-labware-id="<%= presenter.id %>">
+    </div>
+  <% end %>
+
+  <%= card title:'Other Actions', css_class: 'other-actions logged_in_only' do %>
+    <% if presenter.robot? %>
+      <%= simple_state_change_form(presenter) %>
     <% end %>
-  <% else %>
-    <%= simple_state_change_form(presenter) %>
+
+    <% presenter.control_additional_creation do %>
+      <%= render 'creation_dropdown', resource_type: 'plates', resources: presenter.compatible_plate_purposes %>
+      <%= render 'creation_dropdown', resource_type: 'tubes', resources: presenter.compatible_tube_purposes %>
+    <% end %>
+
+    <% presenter.control_library_passing do %>
+      <%= render 'work_completion_form', presenter: presenter %>
+    <% end %>
+
+    <% if presenter.well_failing_applicable? %>
+      <button class='fail-wells'
+        data-toggle="collapse"
+        data-parent="#main-content"
+        href=".well-fail-toggle"
+        aria-controls="well-fail-toggle">Fail Wells</button>
+    <% end %>
+
+    <div id="plate-state">
+      <%= state_change_form(presenter) %>
+    </div>
+
   <% end %>
-<% end %>
-
-<%= card title: 'Adding a Comment', css_class: 'logged_in_only' do %>
-  <div id="asset-comments-add-form" data-comment-title="<%= presenter.comment_title %>"></div>
-<% end %>
-
-<%= card title: 'Adding Custom Metadata', css_class: 'logged_in_only' do %>
-  <div id="labware-custom-metadata-add-form"
-    data-custom-metadata-fields="<%= presenter.custom_metadata_fields %>"
-    data-sequencescape-api="<%= Limber::Application.config.api.v2.connection_options.js_url %>"
-    data-sequencescape-api-key="<%= Limber::Application.config.api.v2.connection_options.authorisation %>"
-    data-sequencescape-url="<%= Limber::Application.config.sequencescape_url %>"
-    data-labware-id="<%= presenter.id %>">
-  </div>
-<% end %>
-
-<%= card title:'Other Actions', css_class: 'other-actions logged_in_only' do %>
-  <% if presenter.robot? %>
-    <%= simple_state_change_form(presenter) %>
-  <% end %>
-
-  <% presenter.control_additional_creation do %>
-    <%= render 'creation_dropdown', resource_type: 'plates', resources: presenter.compatible_plate_purposes %>
-    <%= render 'creation_dropdown', resource_type: 'tubes', resources: presenter.compatible_tube_purposes %>
-  <% end %>
-
-  <% presenter.control_library_passing do %>
-    <%= render 'work_completion_form', presenter: presenter %>
-  <% end %>
-
-  <% if presenter.well_failing_applicable? %>
-    <button class='fail-wells'
-      data-toggle="collapse"
-      data-parent="#main-content"
-      href=".well-fail-toggle"
-      aria-controls="well-fail-toggle">Fail Wells</button>
-  <% end %>
-
-  <div id="plate-state">
-    <%= state_change_form(presenter) %>
-  </div>
-
 <% end %>
