+++ app/views/plates/sidebars/_submission.html.erb
@@ -1,4 +1,7 @@
-<%= render 'pending_submissions', presenter: presenter %>
-<% if presenter.allow_new_submission? %>
-  <%= render 'choose_workflow', presenter: presenter %>
+<%= sidebar do %>
+  <%= render 'logged_out_warning' %>
+    <%= render 'pending_submissions', presenter: presenter %>
+  <% unless presenter.pending_submissions? %>
+    <%= render 'choose_workflow', presenter: presenter %>
+  <% end %>
 <% end %>
