--- app/views/plates/sidebars/_submission_default.html.erb
@@ -1,2 +1,0 @@
-<%= render 'plates/sidebars/submission', presenter: presenter %>
-<%= render 'plates/sidebars/default', presenter: presenter %>
